Testing/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-nested-admin==4.0.2
psycopg2-binary==2.9.5
python-monkey-business==1.0.0
six==1.16.0
sqlparse==0.4.3
